poker-vision-card-snipper/pyproject.toml
[project]
name = "poker-vision-card-snipper"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.9",
]

[tool.uv]
package = false
